requirements.txt
Django==5.0
djangorestframework
djoser
mysqlclient
python-decouple